requirements.txt
# Core dependencies
networkx>=2.8.0
matplotlib>=3.5.0
# pygraphviz>=1.9  # Commented out as it requires graphviz installation
numpy>=1.22.0
colorama>=0.4.4
tabulate>=0.8.9
python-dotenv>=0.19.0
openai>=0.27.0

# Optional dependencies for visualization
# pydot>=1.4.2  # Commented out as it requires graphviz installation

# Testing dependencies
pytest>=7.0.0
